+++ src/StateSmithTest/Input/DrawIo/DrawIoSvgDecodingTests.cs
@@ -61,16 +61,18 @@ public class DrawIoSvgDecodingTests
         action.Should().Throw<DrawIoException>().WithMessage("*opening xml tag `diagram`*found `SomeDiagram`*");
     }
 
+    /// <summary>
+    /// https://github.com/StateSmith/StateSmith/issues/78
+    /// </summary>
     [Fact]
     public void GetDiagramCompressedContents_AdditionalDiagramThrows()
     {
-        Action action = () => DrawIoDecoder.GetDiagramContentsRaw("""
+        DrawIoDecoder.GetDiagramContentsRaw("""
             <mxfile>
                 <diagram id="123" name="Page-1">blah123</diagram>
                 <diagram id="123" name="Page-2">blah123</diagram>
             </mxfile>
             """);
-        action.Should().Throw<DrawIoException>().WithMessage("draw.io files can only contain a single diagram/page for now. See https://github.com/StateSmith/StateSmith/issues/78 .");
     }
 
     [Fact]