+++ src/StateSmithTest/Input/DrawIo/MultiplePagesTest.cs
@@ -1,0 +1,12 @@
+using Xunit;
+
+namespace StateSmithTest.Input.DrawIo;
+
+public class MultiplePagesTest
+{
+    [Fact]
+    public void VertexTest()
+    {
+        TestHelper.CaptureSmRunnerFiles(diagramPath: "MultiplePages.drawio");
+    }
+}